feat: Upgrade local notifications handling
